+++ hirpc/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 
 ## [0.5.2] - 2025-04-12
 
+### Added
+
+- Error payload during handshake.
+
 ### Changed
 
 - Updated hash storage system.